rename config repo
