+++ cmd/configloader/BUILD.bazel
@@ -7,7 +7,7 @@ go_library(
     importpath = "github.com/stebennett/squad-dashboard/cmd/configloader",
     visibility = ["//visibility:private"],
     deps = [
-        "//pkg/jiraconfigrepository",
+        "//pkg/configrepository",
         "//vendor/github.com/Netflix/go-env",
     ],
 )
