+++ pkg/configrepository/BUILD.bazel
@@ -1,0 +1,11 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "configrepository",
+    srcs = [
+        "configrepository.go",
+        "postgresconfigrespository.go",
+    ],
+    importpath = "github.com/stebennett/squad-dashboard/pkg/configrepository",
+    visibility = ["//visibility:public"],
+)
